cleanup and organize facility home components
