+++ src/app/facility/facility-home/facility-home-summary/facility-home-summary.component.ts
@@ -5,7 +5,6 @@ import { UtilityMeterDatadbService } from 'src/app/indexedDB/utilityMeterData-db
 import { IdbAnalysisItem, IdbFacility, IdbUtilityMeter, IdbUtilityMeterData } from 'src/app/models/idb';
 import * as _ from 'lodash';
 import { FacilityHomeService } from '../facility-home.service';
-import { MonthlyAnalysisSummaryData } from 'src/app/models/analysis';
 import { Router } from '@angular/router';
 import { UtilityMeterdbService } from 'src/app/indexedDB/utilityMeter-db.service';
 import { UtilityColors } from 'src/app/shared/utilityColors';
@@ -28,14 +27,6 @@ export class FacilityHomeSummaryComponent implements OnInit {
   sources: Array<MeterSource>;
   naics: string;
   selectedFacilitySub: Subscription;
-  calculatingEnergy: boolean | 'error';
-  calculatingEnergySub: Subscription;
-  calculatingWater: boolean | 'error';
-  calculatingWaterSub: Subscription;
-  monthlyFacilityEnergyAnalysisData: Array<MonthlyAnalysisSummaryData>;
-  monthlyFacilityEnergyAnalysisDataSub: Subscription;
-  monthlyFacilityWaterAnalysisData: Array<MonthlyAnalysisSummaryData>;
-  monthlyFacilityWaterAnalysisDataSub: Subscription;
   constructor(private utilityMeterDataDbService: UtilityMeterDatadbService,
     private facilityDbService: FacilitydbService, private facilityHomeService: FacilityHomeService,
     private router: Router,
@@ -43,34 +34,15 @@ export class FacilityHomeSummaryComponent implements OnInit {
     private exportToExcelTemplateService: ExportToExcelTemplateService) { }
 
   ngOnInit(): void {
-
-    this.calculatingEnergySub = this.facilityHomeService.calculatingEnergy.subscribe(val => {
-      this.calculatingEnergy = val;
-    });
-    this.calculatingWaterSub = this.facilityHomeService.calculatingWater.subscribe(val => {
-      this.calculatingWater = val;
-    });
-
     this.selectedFacilitySub = this.facilityDbService.selectedFacility.subscribe(val => {
       this.facility = this.facilityDbService.selectedFacility.getValue();
       this.setNAICS();
       this.setFacilityStatus();
     });
-    this.monthlyFacilityEnergyAnalysisDataSub = this.facilityHomeService.monthlyFacilityEnergyAnalysisData.subscribe(val => {
-      this.monthlyFacilityEnergyAnalysisData = val;
-    });
-    this.monthlyFacilityWaterAnalysisDataSub = this.facilityHomeService.monthlyFacilityWaterAnalysisData.subscribe(val => {
-      this.monthlyFacilityWaterAnalysisData = val;
-    });
-
   }
 
   ngOnDestroy() {
-    this.monthlyFacilityEnergyAnalysisDataSub.unsubscribe();
-    this.monthlyFacilityWaterAnalysisDataSub.unsubscribe();
     this.selectedFacilitySub.unsubscribe();
-    this.calculatingEnergySub.unsubscribe();
-    this.calculatingWaterSub.unsubscribe();
   }
 
   navigateTo(urlStr: string) {
