--- src/app/facility/facility-home/meter-card/meter-card.component.css
@@ -1,11 +1,0 @@
-.card-header{
-    background-color: white;
-}
-
-.badge{
-    font-size: 12px;
-}
-
-h3 a {
-    font-size: 24.5px;
-}
