--- src/app/facility/facility-home/scorecard-status/scorecard-status.component.css
@@ -1,11 +1,0 @@
-.card-header, thead tr{
-    background-color: white;
-}
-
-th, td{
-    text-align: center;
-}
-
-.fa-check-circle{
-    color: green;
-}
