--- src/app/facility/facility-home/scorecard-status/scorecard-status.component.html
@@ -1,69 +1,0 @@
-<div class="card">
-    <!-- <div class="card-header">
-        <h3 class="card-title mb-0">Energy Savings Projects</h3>
-    </div>
-    <div class="card-body"> -->
-    <table class="table">
-        <thead>
-            <tr>
-                <th>
-                    Savings Project
-                </th>
-                <th>
-                    Scorecard
-                </th>
-                <th>
-                    Implemented
-                </th>
-                <th>
-                    Energy Savings
-                </th>
-            </tr>
-        </thead>
-        <tbody>
-            <tr>
-                <td>
-                    Change lighting
-                </td>
-                <td>
-                    <a>2018</a>
-                </td>
-                <td class="text-center">
-                    {{dateOne | date:'MMM, y'}}
-                </td>
-                <td>
-                    200,000 MMBtu/yr
-                </td>
-            </tr>
-            <tr>
-                <td>
-                    Replace Boiler
-                </td>
-                <td>
-                    <a>2018</a>
-                </td>
-                <td class="text-center">
-                    {{dateTwo | date:'MMM, y'}}
-                </td>
-                <td>
-                    450,000 MMBtu/yr
-                </td>
-            </tr>
-            <tr>
-                <td>
-                    Install Solar Panels
-                </td>
-                <td>
-                    <a>2022</a>
-                </td>
-                <td class="text-center">
-                    &mdash;
-                </td>
-                <td>
-                    150,000 MMBtu/yr
-                </td>
-            </tr>
-        </tbody>
-    </table>
-    <!-- </div> -->
-</div>
